backend/requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
celery==5.4.0
redis==5.2.1
dj-database-url==2.3.0
django-cors-headers==4.6.0
drf-spectacular==0.29.0
Pillow==11.0.0
django-auditlog==3.0.0
